+++ src/commands/index.ts
@@ -1,4 +1,5 @@
 // start:ng42.barrel
-export * from './goto-parent';
+export * from "./goto-parent";
 // end:ng42.barrel
-
+export * from "./logger";
+export * from "./create-node-map";