+++ src/providers/symbols.ts
@@ -1,22 +1,22 @@
 "use strict";
 
 import {
+  commands,
+  Position,
   SymbolInformation,
+  SymbolKind,
   TextDocument,
   Uri,
-  commands,
-  workspace,
-  Position,
-  SymbolKind
+  workspace
 } from "vscode";
-import { hasParents, baseClassRegex, getAllDefinitions } from ".";
-import { log } from "../commands/logger";
+import { baseClassRegex, getAllDefinitions, hasParents } from ".";
+import { log, createNamedNodeLookUp } from "../commands";
 
 /**
  *
  * Get all symbols of the document with given uri.
  * 1. open the target document
- * 2. execute symbol provider against the just opened target document
+ * 2. excute symbol provider against the just opened target document
  *
  * Note: the function open file in the background. it will show the document to the user.
  *
@@ -52,7 +52,41 @@ export async function getSymbolsOpenedUri(
     throw error;
   }
 }
-
+// export function getCSharpSymbols(
+//   doc: TextDocument,
+//   classSymbol: SymbolInformation,
+//   symbols: SymbolInformation[]
+// ): SymbolInformation[] {
+//   let parentsAndInterfaces = doc.getText();
+//   const classIndex = parentsAndInterfaces.indexOf("class " + classSymbol.name);
+//   if (classIndex === -1) {
+//     return [];
+//   }
+//   parentsAndInterfaces = parentsAndInterfaces.slice(classIndex);
+//   const columnIndex = parentsAndInterfaces.indexOf(":");
+//   let parents: string[] = [];
+//   if (columnIndex >= 0) {
+//     let parentText = parentsAndInterfaces.slice(columnIndex + 1);
+//     parents = parentText.slice(0, parentText.indexOf("{")).split(",");
+//     if (parents) {
+//       parents = parents.map(i => i.trim());
+//     }
+//   } else {
+//     return [];
+//   }
+//   // get interface symbols by names from given symbols list.
+//   let parentSymbols: SymbolInformation[] = [];
+//   parents.forEach(i => {
+//     const s = symbols.find(
+//       // symbols often doesn't include generic part, remove it to find all valid interface symbols
+//       s => s.name.replace(/(<).+(>)/, "") === i.replace(/(<).+(>)/, "")
+//     );
+//     if (s) {
+//       parentSymbols.push(s);
+//     }
+//   });
+//   return parentSymbols;
+// }
 /**
  *
  * Return symbols of interfaces of a given class.
@@ -62,24 +96,70 @@ export async function getSymbolsOpenedUri(
  * @param symbols symbols of current document.
  * @returns symbols of interfaces of given class.
  */
-export function getInterfaceSymbols(
+export async function getInterfaceSymbols(
   doc: TextDocument,
   classSymbol: SymbolInformation,
   symbols: SymbolInformation[]
-): SymbolInformation[] {
+): Promise<ParentSymbol[]> {
   const interfaces = getInterfaceNames(doc, classSymbol);
   // get interface symbols by names from given symbols list.
-  let interfaceSymbols: SymbolInformation[] = [];
-  interfaces.forEach(i => {
-    const s = symbols.filter(
-      // symbols often doesn't include generic part, remove it to find all valid interface symbols
-      s => s.name.replace(/(<).+(>)/, "") === i.replace(/(<).+(>)/, "")
-    );
-    if (s && s.length > 0) {
-      interfaceSymbols.push(s[0]);
+  let interfaceSymbols: ParentSymbol[] = [];
+  for (let i of interfaces) {
+    let s = await getParentSymbol(i, doc, symbols);
+    if (s) {
+      interfaceSymbols.push(s);
     }
-  });
+  }
   return interfaceSymbols;
+}
+
+export async function getParentSymbol(
+  name: string,
+  doc: TextDocument,
+  symbols: SymbolInformation[]
+): Promise<ParentSymbol | undefined> {
+  name = name.replace(/(<).+(>)/, "");
+  const symbol = symbols.find(s => s.name === name);
+  if (symbol) {
+    return {
+      name: name,
+      position: new Position(
+        symbol.location.range.start.line,
+        symbol.location.range.start.character + 1
+      )
+    };
+  }
+  const lookup = await createNamedNodeLookUp(doc.getText());
+  const node = lookup(name);
+  if (!node) {
+    log(name + " node not found");
+    return;
+  }
+  const p = doc.positionAt(node.pos);
+
+  return { name: name, position: new Position(p.line, p.character + 1) };
+}
+
+/**
+ * Return symbols of base class of a given class.
+ *
+ * @param doc current document
+ * @param classSymbol symbol of class to look for base class for.
+ * @param symbols symbols of current document.
+ *
+ * @returns Symbol of base class of the given class.
+ */
+export async function getBaseClassSymbol(
+  doc: TextDocument,
+  classSymbol: SymbolInformation,
+  symbols: SymbolInformation[]
+): Promise<ParentSymbol | undefined> {
+  let parentClassName = getBaseClassName(doc, classSymbol);
+
+  if (!parentClassName) {
+    return;
+  }
+  return await getParentSymbol(parentClassName, doc, symbols);
 }
 
 /**
@@ -119,30 +199,29 @@ export function getInterfaceNames(
   }
   return interfaces;
 }
-
 /**
- * Return symbols of base class of a given class.
+ * get class's declaration text.
+ * Note: it is only contains the given class's body, not the whole body.
+ * when class is changed by the user, only the first part to the changing point of the class body will be returned.
+ * Therefore, changing class may interrupt classIO until class text returned contains the sholw declaration of the class.
+ * @param doc : current text document
+ * @param classSymbol symbol of target class to look for.
  *
- * @param doc current document
- * @param classSymbol symbol of class to look for base class for.
- * @param symbols symbols of current document.
- *
- * @returns Symbol of base class of the given class.
+ * @returns string of class's body starting from the end of class name to the end of class.
  */
-export function getBaseClassSymbol(
+function getClassTextFromClassName(
   doc: TextDocument,
-  classSymbol: SymbolInformation,
-  symbols: SymbolInformation[]
-): SymbolInformation | undefined {
-  let parentClassName = getBaseClassName(doc, classSymbol);
-  if (!parentClassName) {
-    return;
+  classSymbol: SymbolInformation
+): string {
+  const classText = doc.getText(classSymbol.location.range);
+  const classIndex = classText.indexOf("class " + classSymbol.name);
+  if (classIndex === -1) {
+    return "";
+  } else {
+    return classText.slice(
+      classIndex + classSymbol.name.length + "class ".length
+    );
   }
-  return symbols.find(
-    // remove generic signature.
-    s =>
-      s.name.replace(/(<).+(>)/, "") === parentClassName.replace(/(<).+(>)/, "")
-  );
 }
 
 export function getBaseClassName(
@@ -178,29 +257,9 @@ export function getBaseClassName(
   return parentClassName;
 }
 
-/**
- * get class's declaration text.
- * Note: it is only contains the given class's body, not the whole body.
- * when class is changed by the user, only the first part to the changing point of the class body will be returned.
- * Therefore, changing class may interrupt classlens until class text returned contains the sholw declaration of the class.
- * @param doc : current text document
- * @param classSymbol symbol of target class to look for.
- *
- * @returns string of class's body starting from the end of class name to the end of class.
- */
-function getClassTextFromClassName(
-  doc: TextDocument,
-  classSymbol: SymbolInformation
-): string {
-  const classText = doc.getText(classSymbol.location.range);
-  const classIndex = classText.indexOf("class " + classSymbol.name);
-  if (classIndex === -1) {
-    return "";
-  } else {
-    return classText.slice(
-      classIndex + classSymbol.name.length + "class ".length
-    );
-  }
+export interface ParentSymbol {
+  name: string;
+  position: Position;
 }
 
 export function getNameSpacePosition(