+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.7.0
+
+Now works for vscode 1.2.4.
+
+Now works for namespace with types in tsconfig.json(better support for react projects).
+
 ## 0.6.2
 
 Add limitations in readme.
@@ -45,9 +51,9 @@ Optimise cache and performance increased
 
 ## 0.1.0
 
-* Add Javascript support
-* fix bug: child and parent class in the same file now should work.
+- Add Javascript support
+- fix bug: child and parent class in the same file now should work.
 
 ## 0.0.1
 
-* Initial release
+- Initial release